title: remove preceding number prefix
